requirements.txt

streamlit>=1.18.1
rembg==2.0.28
onnxruntime>=1.15.1
pillow>=9.0.0
numpy>=1.23.0
